refactor project structure, separate project routes from index.js
