+++ src/routes/projects.js
@@ -1,30 +1,28 @@
 const express = require('express')
 const { Validator } = require('express-json-validator-middleware')
-const db = require('./db')
-const { projectSchema } = require('./projectSchema')
+const db = require('../data/db')
+const { projectSchema } = require('../schemas/projectSchema')
 
-const port = process.env.PORT || 3000
-const app = express()
+// use a router instead of express() so that we are using the same instance from index.js
+const router = express.Router()
 const { validate } = new Validator()
 
-app.use(express.json())
 
-app.get('/projects', (req, res) => {
+router.get('/', (req, res) => {
     // no need to add response code in express bc it will do so automatically
     res.json(db.projects)
 })
 
-app.get('/projects/:id', (req, res) => {
+router.get('/:id', (req, res) => {
     // :id can be retrieved using req.params object
     const project = db.projects.find((proj) => proj.id == req.params.id)
     res.json(project)
 })
 
-app.post(
-    '/projects',
+router.post(
+    '/',
     validate({ body: projectSchema }),
     (req, res) => {
-    // expect new project to be passed in via the request body
     try {
         const newProject = req.body
         console.log('new project is', newProject)
@@ -36,7 +34,7 @@ app.post(
     }
 })
 
-app.delete('/projects/:id', (req, res) => {
+router.delete('/:id', (req, res) => {
     const projectToDelete = db.projects.find((proj) => proj.id == req.params.id)
     if(!projectToDelete) return res.status(400).send('There is no matching project with id', req.params.id)
     //TODO implement delete functionality once it is connected to a database
@@ -45,8 +43,5 @@ app.delete('/projects/:id', (req, res) => {
     res.send(`successfully removed project ${JSON.stringify(projectToDelete)}`)
 })
 
+module.exports = router
 
-// we must add a listener so that the response is being listened when we run the server
-app.listen(port, () => {
-    console.log(`projects api now listening on port ${port}`)
-})
